+++ pytsite/reload/_api.py
@@ -1,7 +1,7 @@
 """PytSite Reload API Functions.
 """
-from os import path as _path, utime as _utime
-from pytsite import reg as _reg, logger as _logger, events as _events, console as _console, lang as _lang
+from os import path as _path
+from pytsite import reg as _reg, events as _events, console as _console, lang as _lang, util as _util
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -11,25 +11,19 @@ RELOAD_MSG_ID = 'pytsite.reload@reload_required'
 _flag = False
 
 
-def reload(console_notify: bool = True):
-    """Touch 'touch.reload' file.
+def reload():
+    """Modify 'touch.reload' file.
     """
     touch_reload_path = _path.join(_reg.get('paths.storage'), 'touch.reload')
 
     _events.fire('pytsite.reload.before')
 
-    if not _path.exists(touch_reload_path):
-        with open(touch_reload_path, 'w'):
-            pass
-    else:
-        _utime(touch_reload_path, None)
+    with open(touch_reload_path, 'wt') as f:
+        f.write(_util.w3c_datetime_str())
 
     _events.fire('pytsite.reload')
 
-    _logger.info('Application is reloading')
-
-    if console_notify:
-        _console.print_info(_lang.t('pytsite.reload@app_is_reloading'))
+    _console.print_info(_lang.t('pytsite.reload@app_is_reloading'))
 
 
 def set_flag():